Add files via upload
This team focused on simple, uncomplicated solutions, and we created two meetings. In the first, we distributed roles, and in the second, we discussed tasks. The UI team put a lot of effort into research and implementation. The JS team focused on having an organized structure, which was done with the help of sources and modifications to the code. Engineer: Faisal, in the beginning, and this was done. Edited by Engineer: Zayid
And the engineer: Adham